Properly emulate 256 colors + export fg and bg color
We pretend to be xterm we have to be able to support 256 colors.

libvterm does not support xterm specific CSI that allows applications to
request fg and bg colors (which is used for example by vim to determine
if dark or light color scheme should be used) so we have to propagate
that via env variable.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>